Drawer: rename drawer.rendering.strategy
