+++ packages/devextreme/js/__internal/ui/drawer/drawer.rendering.strategy.ts
